f5-tts-app/src/utils/model_utils.py

import torch
import os
import yaml
import sys

def load_model_and_vocoder(model_path, config):
    """
    Load TTS model and vocoder
    
    Args:
        model_path: Path to model checkpoint
        config: Configuration dictionary
        
    Returns:
        model, vocoder: Loaded models
    """
    print(f"Would load model from {model_path} with config: {config['model']['name']}")
    
    # Here we would normally load the actual model, but for testing we'll create placeholders
    class DummyModel:
        def __init__(self, config):
            self.config = config
            self.name = config['model']['name']
        
        def eval(self):
            return self
            
        def to(self, device):
            return self
    
    class DummyVocoder:
        def __init__(self):
            pass
            
        def eval(self):
            return self
            
        def to(self, device):
            return self
    
    model = DummyModel(config)
    vocoder = DummyVocoder()
    
    return model, vocoder
