f5-tts-app/requirements.txt
torch>=1.10.0
torchaudio>=2.0.0
safetensors
numpy>=1.20.0
einops
transformers
pypinyin
flask>=2.0.0
flask-cors>=3.0.10
pyyaml>=6.0
scipy>=1.7.0
librosa>=0.9.0
huggingface_hub